Adds the Ceph-Admin relation and kubernetes pv creation using ceph-rbd
This is a feature branch commit rollup of:

- Adds the ceph-storage relation
- Revert "The first attempt at implementing the ceph-admin interface."
- Adds state boundaries for controlling ceph-storage
- Adds debug statements, as well as a health check for ceph
