+++ cluster/juju/layers/kubernetes-master/actions.yaml
@@ -1,2 +1,22 @@
 restart-master:
   description: restart the apiserver, controller-manager, and scheduler
+create-rbd-pv:
+    description: Create persistant storage when related to ceph
+    options:
+      name:
+        type: string
+        default: ""
+        description: Name the persistent volume
+      size:
+        type: int
+        default: 0
+        description: Size in mb of the rbd volume
+      mode:
+        type: string
+        default: ReadWriteOnce
+        description: Access mode for the persistent volume.
+      filesystem:
+        type: string
+        default: xfs
+        description: FS type to format the volume
+
