+++ cluster/juju/layers/kubernetes-master/layer.yaml
@@ -6,6 +6,7 @@ includes:
   - 'interface:http'
   - 'interface:sdn-plugin'
   - 'interface:kube-dns'
+  - 'interface:ceph-admin'
   - 'interface:public-address'
 options:
   tls-client:
@@ -14,3 +15,4 @@ options:
     server_key_path: '/srv/kubernetes/server.key'
     client_certificate_path: '/srv/kubernetes/client.crt'
     client_key_path: '/srv/kubernetes/client.key'
+
