+++ cluster/juju/layers/kubernetes-master/templates/ceph-secret.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: ceph-secret
+type: Opaque
+data:
+  key: {{ secret }}
