+++ cluster/juju/layers/kubernetes-master/templates/ceph.conf
