+++ cluster/juju/layers/kubernetes-master/templates/rbd-persistent-volume.yaml
@@ -1,0 +1,26 @@
+# JUJU Internal Template used to enlist RBD volumes from the
+# `create-rbd-pv` action. This is a temporary file on disk to enlist resources.
+apiVersion: v1
+kind: PersistentVolume
+metadata:
+  name: {{ RBD_NAME }}
+  annotations:
+    volume.beta.kubernetes.io/storage-class: "rbd"
+spec:
+  capacity:
+    storage: {{ RBD_SIZE }}M
+  accessModes:
+    - {{ PV_MODE }}
+  rbd:
+    monitors:
+      {% for host in monitors %}
+      - {{ host }}
+      {% endfor %}
+    pool: rbd
+    image: {{ RBD_NAME }}
+    user: admin
+    secretRef:
+      name: ceph-secret
+    fsType: {{ RBD_FS }}
+    readOnly: false
+    # persistentVolumeReclaimPolicy: Recycle
