+++ cluster/juju/layers/kubernetes-worker/layer.yaml
@@ -3,6 +3,7 @@ includes:
   - 'layer:docker'
   - 'layer:tls-client'
   - 'interface:http'
+  - 'interface:sdn-plugin'
   - 'interface:kube-dns'
 repo: https://github.com/kubernetes/kubernetes
 options:
